+++ src/tree/v7400.rs
@@ -36,11 +36,15 @@ pub struct Tree {
 
 impl Tree {
     /// Returns the root node.
+    #[inline]
+    #[must_use]
     pub fn root(&self) -> NodeHandle<'_> {
         NodeHandle::new(self, self.root_id)
     }
 
     /// Creates a new `Tree`.
+    #[inline]
+    #[must_use]
     fn new(
         arena: Arena<NodeData>,
         node_names: StringInterner<DefaultBackend<NodeNameSym>>,
@@ -58,6 +62,7 @@ impl Tree {
     /// # Panics
     ///
     /// Panics if a node with the given node ID does not exist in the tree.
+    #[must_use]
     pub(crate) fn node(&self, node_id: NodeId) -> &indextree::Node<NodeData> {
         self.arena.get(node_id.raw()).unwrap_or_else(|| {
             panic!(
@@ -72,6 +77,7 @@ impl Tree {
     /// # Panics
     ///
     /// Panics if the given symbol is not used in the tree.
+    #[must_use]
     pub(crate) fn resolve_node_name(&self, sym: NodeNameSym) -> &str {
         self.node_names
             .resolve(sym)
@@ -79,11 +85,13 @@ impl Tree {
     }
 
     /// Returns node name symbol if available.
+    #[must_use]
     pub(crate) fn node_name_sym(&self, name: &str) -> Option<NodeNameSym> {
         self.node_names.get(name)
     }
 
     /// Checks whether or not the given node ID is used in the tree.
+    #[must_use]
     pub(crate) fn contains_node(&self, node_id: NodeId) -> bool {
         self.arena.get(node_id.raw()).is_some()
     }
@@ -162,6 +170,42 @@ impl Tree {
         node.get_mut().append_attribute(v)
     }
 
+    /// Returns a mutable reference to the node attribute at the given index.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the given node ID is invalid (i.e. not used or root node).
+    #[must_use]
+    pub fn get_attribute_mut(&mut self, node_id: NodeId, i: usize) -> Option<&mut AttributeValue> {
+        let node = self.arena.get_mut(node_id.raw()).expect("Invalid node ID");
+        node.get_mut().get_attribute_mut(i)
+    }
+
+    /// Takes all attributes as a `Vec`.
+    ///
+    /// After calling this, the node will have no attributes (until other values are set).
+    ///
+    /// # Panics
+    ///
+    /// Panics if the given node ID is invalid (i.e. not used or root node).
+    pub fn take_attributes_vec(&mut self, node_id: NodeId) -> Vec<AttributeValue> {
+        let node = self.arena.get_mut(node_id.raw()).expect("Invalid node ID");
+        node.get_mut().replace_attributes(Default::default())
+    }
+
+    /// Sets the given `Vec` of attribute values as the node attributes.
+    ///
+    /// After calling this, the node will have only the given attributes.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the given node ID is invalid (i.e. not used or root node).
+    pub fn set_attributes_vec(&mut self, node_id: NodeId, new: Vec<AttributeValue>) {
+        let node = self.arena.get_mut(node_id.raw()).expect("Invalid node ID");
+        // Ignore the returned value.
+        node.get_mut().replace_attributes(new);
+    }
+
     /// Compares trees strictly.
     ///
     /// Returns `true` if the two trees are same.
@@ -170,6 +214,8 @@ impl Tree {
     ///
     /// Note that this method compares tree data, not internal states of the
     /// objects.
+    #[inline]
+    #[must_use]
     pub fn strict_eq(&self, other: &Self) -> bool {
         self.root().strict_eq(&other.root())
     }
@@ -177,6 +223,8 @@ impl Tree {
     /// Pretty-print the tree for debugging purpose.
     ///
     /// Be careful, this output format may change in future.
+    #[inline]
+    #[must_use]
     pub fn debug_tree(&self) -> impl fmt::Debug + '_ {
         DebugTree { tree: self }
     }
@@ -204,6 +252,7 @@ struct DebugTree<'a> {
 }
 
 impl fmt::Debug for DebugTree<'_> {
+    #[inline]
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         let v = DebugNodeHandle {
             node: self.tree.root(),
@@ -243,5 +292,239 @@ impl fmt::Debug for DebugNodeHandleChildren<'_> {
                     .map(|child| DebugNodeHandle { node: child }),
             )
             .finish()
+    }
+}
+
+/// Event of depth-first traversal.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum DepthFirstTraversed {
+    /// Opening of a node.
+    Open(NodeId),
+    /// Closing of a node.
+    Close(NodeId),
+}
+
+impl DepthFirstTraversed {
+    /// Returns the node ID.
+    #[inline]
+    #[must_use]
+    pub fn node_id(self) -> NodeId {
+        match self {
+            Self::Open(id) => id,
+            Self::Close(id) => id,
+        }
+    }
+
+    /// Returns true if the event is node open.
+    #[inline]
+    #[must_use]
+    pub fn is_open(self) -> bool {
+        matches!(self, Self::Open(_))
+    }
+
+    /// Returns true if the event is node close.
+    #[inline]
+    #[must_use]
+    pub fn is_close(self) -> bool {
+        matches!(self, Self::Close(_))
+    }
+
+    /// Returns the opened node ID.
+    #[inline]
+    #[must_use]
+    pub fn node_id_open(self) -> Option<NodeId> {
+        match self {
+            Self::Open(id) => Some(id),
+            Self::Close(_) => None,
+        }
+    }
+
+    /// Returns the closed node ID.
+    #[inline]
+    #[must_use]
+    pub fn node_id_close(self) -> Option<NodeId> {
+        match self {
+            Self::Open(_) => None,
+            Self::Close(id) => Some(id),
+        }
+    }
+
+    /// Returns next (forward) event.
+    ///
+    /// Returns `None` for `Close(root_id)`.
+    #[must_use]
+    pub fn next(self, tree: &Tree) -> Option<Self> {
+        let next = match self {
+            Self::Open(current) => {
+                // Dive into the first child if available, or otherwise leave the node.
+                match current.to_handle(tree).first_child() {
+                    Some(child) => Self::Open(child.node_id()),
+                    None => Self::Close(current),
+                }
+            }
+            Self::Close(current) => {
+                // Dive into the next sibling if available, or leave the parent.
+                let node = current.to_handle(tree);
+                match node.next_sibling() {
+                    Some(next_sib) => Self::Open(next_sib.node_id()),
+                    None => Self::Close(node.parent()?.node_id()),
+                }
+            }
+        };
+        Some(next)
+    }
+
+    /// Returns previous (backward next) event.
+    ///
+    /// Note that this backward traversal returns `Clone` first, and `Open`
+    /// later for every node.
+    ///
+    /// Returns `None` for `Open(root_id)`.
+    #[must_use]
+    pub fn prev(self, tree: &Tree) -> Option<Self> {
+        let prev = match self {
+            Self::Close(current) => {
+                // Dive into the last child if available, or otherwise leave the node.
+                match current.to_handle(tree).last_child() {
+                    Some(child) => Self::Close(child.node_id()),
+                    None => Self::Open(current),
+                }
+            }
+            Self::Open(current) => {
+                // Dive into the previous sibling if available, or leave the parent.
+                let node = current.to_handle(tree);
+                match node.previous_sibling() {
+                    Some(prev_sib) => Self::Close(prev_sib.node_id()),
+                    None => Self::Open(node.parent()?.node_id()),
+                }
+            }
+        };
+        Some(prev)
+    }
+}
+
+/// A type to traverse a node and its descendants in depth-first order.
+///
+/// This type has two cursors, forward cursor and backward cursor.
+/// In the initial state, forward cursor points to the opening of the root node,
+/// and the backward cursor points to the ending of the root node.
+///
+/// Forward cursor advances forward, and backward cursor advances backward.
+/// `next_forward` and `next_backward` methods returns the node ID the
+/// corresponding cursor points to, and advances the cursor.
+///
+/// When the forward cursor goes after the backward cursor, then all events are
+/// considered emitted.
+#[derive(Debug, Clone, Copy)]
+pub struct DepthFirstTraverseSubtree {
+    /// Next (forward and backward) events to return.
+    cursors: Option<(DepthFirstTraversed, DepthFirstTraversed)>,
+}
+
+impl DepthFirstTraverseSubtree {
+    /// Creates a new object.
+    #[inline]
+    #[must_use]
+    fn with_root_id(root: NodeId) -> Self {
+        Self {
+            cursors: Some((
+                DepthFirstTraversed::Open(root),
+                DepthFirstTraversed::Close(root),
+            )),
+        }
+    }
+
+    /// Returns the forward next traversal event and advances the forward cursor.
+    #[inline]
+    pub fn next_forward(&mut self, tree: &Tree) -> Option<DepthFirstTraversed> {
+        let (forward, backward) = self.cursors?;
+        if forward == backward {
+            self.cursors = None;
+        } else {
+            let next_of_next = forward
+                .next(tree)
+                .expect("`forward` should point before `backward`");
+            self.cursors = Some((next_of_next, backward));
+        }
+        Some(forward)
+    }
+
+    /// Returns the backward next traversal event and advances the backward cursor.
+    #[inline]
+    pub fn next_backward(&mut self, tree: &Tree) -> Option<DepthFirstTraversed> {
+        let (forward, backward) = self.cursors?;
+        if forward == backward {
+            self.cursors = None;
+        } else {
+            let next_of_next = backward
+                .prev(tree)
+                .expect("`forward` should point before `backward`");
+            self.cursors = Some((forward, next_of_next));
+        }
+        Some(backward)
+    }
+
+    /// Returns the forward next traversal event without advancing the cursor.
+    #[inline]
+    #[must_use]
+    pub fn peek_forward(&self) -> Option<DepthFirstTraversed> {
+        self.cursors.map(|(forward, _backward)| forward)
+    }
+
+    /// Returns the backward next traversal event without advancing the cursor.
+    #[inline]
+    #[must_use]
+    pub fn peek_backward(&self) -> Option<DepthFirstTraversed> {
+        self.cursors.map(|(_forward, backward)| backward)
+    }
+
+    /// Returns the forward next `Open` traversal event and advances the forward cursor.
+    ///
+    /// This makes it easy to forward-traverse the subtree in preorder.
+    pub fn next_open_forward(&mut self, tree: &Tree) -> Option<NodeId> {
+        loop {
+            let next = self.next_forward(tree)?;
+            if let DepthFirstTraversed::Open(id) = next {
+                return Some(id);
+            }
+        }
+    }
+
+    /// Returns the forward next `Close` traversal event and advances the forward cursor.
+    ///
+    /// This makes it easy to forward-traverse the subtree in postorder.
+    pub fn next_close_forward(&mut self, tree: &Tree) -> Option<NodeId> {
+        loop {
+            let next = self.next_forward(tree)?;
+            if let DepthFirstTraversed::Close(id) = next {
+                return Some(id);
+            }
+        }
+    }
+
+    /// Returns the backward next `Open` traversal event and advances the forward cursor.
+    ///
+    /// This makes it easy to backward-traverse the subtree in postorder.
+    #[must_use]
+    pub fn next_open_backward(&mut self, tree: &Tree) -> Option<NodeId> {
+        loop {
+            let next = self.next_backward(tree)?;
+            if let DepthFirstTraversed::Open(id) = next {
+                return Some(id);
+            }
+        }
+    }
+
+    /// Returns the backward next `Close` traversal event and advances the forward cursor.
+    ///
+    /// This makes it easy to backward-traverse the subtree in preorder.
+    #[must_use]
+    pub fn next_close_backward(&mut self, tree: &Tree) -> Option<NodeId> {
+        loop {
+            let next = self.next_backward(tree)?;
+            if let DepthFirstTraversed::Close(id) = next {
+                return Some(id);
+            }
+        }
     }
 }